# Hierarchies: "Hide if no children" performance improvement