--- a/apps/performance-tests/src/main.ts
+++ b/apps/performance-tests/src/main.ts
@@ -7,7 +7,7 @@ import { IModelHost } from "@itwin/core-backend";
 import { Logger, LogLevel } from "@itwin/core-bentley";
 import { createLogger } from "@itwin/presentation-core-interop";
 import { setLogger } from "@itwin/presentation-hierarchies";
-import { Datasets } from "./Datasets";
+import { Datasets } from "./util/Datasets";
 
 before(async () => {
   Logger.initializeToConsole();