At some point we stopped passing the `stopOnFirstChild` argument to `createHideIfNoChildrenOperator`, which made children checks for nodes, having many children with "hide if no children" flag, very slow.

Added a performance test to avoid future regressions. Results on my local machine before the change:
```
PASS   | hide if no children required to finalize root, w/o children | 33829.58 ms | count: 40, max: 153.00, p95: 111.00, median: 59.50
PASS   | hide if no children required to finalize root, w/ children  | 16171.97 ms | count: 33, max: 164.00, p95: 114.00, median: 36.00
```
After the change:
```
PASS   | hide if no children required to finalize root, w/o children | 33336.39 ms | count: 2, max: 34.00, p95: 34.00, median: 28.50
PASS   | hide if no children required to finalize root, w/ children  | 107.39 ms   | count: 1, max: 22.00, p95: 22.00, median: 22.00
```